MM-1159 extract Omnigent resource harvester
